correlated energy densities in loc_analyzers, oep module with more stable algorithm in FPXCPotentialAnalyzer
